fix: getHash function now get uint48 for validUntil, validAfter

+++ contracts/samples/VerifyingPaymaster.sol
@@ -55,7 +55,7 @@ contract VerifyingPaymaster is BasePaymaster {
      * note that this signature covers all fields of the UserOperation, except the "paymasterAndData",
      * which will carry the signature itself.
      */
-    function getHash(UserOperation calldata userOp, uint64 validUntil, uint64 validAfter)
+    function getHash(UserOperation calldata userOp, uint48 validUntil, uint48 validAfter)
     public view returns (bytes32) {
         //can't use userOp.hash(), since it contains also the paymasterAndData itself.
